% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sce}
\alias{check_sce}
\title{Check if an object is a Summarized- or SingleCellExperiment and that it contains a certain assay}
\usage{
check_sce(x, assay)
}
\description{
Check if an object is a Summarized- or SingleCellExperiment and that it contains a certain assay
}
\keyword{internal}
