% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circtools_060226.R
\name{circ.confint}
\alias{circ.confint}
\title{Calculate circular confidence interval}
\usage{
circ.confint(
  x,
  w = NULL,
  conf.level = 0.95,
  axial = FALSE,
  na.rm = TRUE,
  method = c("ori", "clt"),
  bias = TRUE
)
}
\arguments{
\item{x}{Numeric vector of angles (in degrees)}

\item{w}{Optional weights}

\item{conf.level}{Confidence level (e.g. 0.95)}

\item{axial}{Logical. if TRUE treat data as axial (0–180°)}

\item{na.rm}{Logical. if TRUE remove NA pairs}

\item{method}{“clt” or “ori”. Calculate the confidence interval based on the central limit theorem ("clt", standard) or the small-angle formula ("ori") that Oriana uses. "ori" does not use the concentration parameter k, and instead assumes that the data are tightly clustered. "ori" can work for low n-sizes that are highly clustered.}

\item{bias}{Bias correction of the concentration parameter for low sample sizes, usually n<30. Only applicable when method = "clt".}
}
\description{
Calculates the circular confidence intervals. Accepts weighted, and axial data.
}
