% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circtoolsstats_011025.R
\name{circ.plot}
\alias{circ.plot}
\title{Plot circular data}
\usage{
circ.plot(
  x,
  r = NULL,
  weight = TRUE,
  ci = 0.95,
  labels = c("geo", "rad", "deg"),
  cex = 1,
  sep = 0.05,
  pch = 16,
  mvwidth = 0.2,
  mvlwd = 3,
  arrowwidth = 0.1,
  arrowlwd = 0.1,
  shrink = 1.2,
  zero = pi/2,
  stack = TRUE,
  rotation = "clock",
  col = "gray40",
  mv = TRUE,
  mvcol = "blue",
  cicol = "black",
  bins = 120,
  cilwd = 2,
  main = NULL,
  cidist = 0.3,
  citick = 0.05,
  axial = FALSE,
  units = c("deg", "rad"),
  norm = c("sum", "n")
)
}
\arguments{
\item{x}{A vector of angles in degrees.}

\item{r}{A vector of r-values (vector lengths).}

\item{weight}{Logical. Default: TRUE, to weight your angles by r-values.}

\item{ci}{The confidence level to calculate. Default: 0.95. ci=NULL to not plot confidence interval.}

\item{labels}{Plot your data with geographical or numeric labels. c("geo", "deg", "rad").}

\item{sep}{Distance between dots.}

\item{pch}{Point charactter to use.}

\item{mvwidth}{Width of mean vector.}

\item{mvlwd}{Line width of mean vector.}

\item{arrowwidth}{Width of individual vectors}

\item{arrowlwd}{Line width of individual vectors}

\item{shrink}{Parameter that controls the size of the plotted circle. Default is 1.2.}

\item{zero}{Location of Zero. Default = pi/2.}

\item{stack}{Logical. To stack datapoints or not}

\item{rotation}{Rotation of the circle. c("clock", "counter"). Default: "clock"}

\item{col}{Color of the points and individual vectors.}

\item{mv}{Logical. To plot mean vector or not.}

\item{mvcol}{Color of mean vector}

\item{cicol}{Color of confidence interval}

\item{bins}{Number of bins to partition the circle}

\item{cilwd}{Linewidth of Confidence interval}

\item{main}{Plot title}

\item{cidist}{Distance from circle to confidence interval. Default: 0.2}

\item{citick}{Width of confidence interval edges}

\item{axial}{Logical. Default: FALSE.}

\item{norm}{Method for normalization of vector length. "sum" = sum(weights)^2 / sum(weights^2). "n" = number of observations}
}
\value{
A circular plot
}
\description{
Easy plotting of circular data through the 'circular' package.
}
\examples{
x <- rad2deg(rvonmises(50, pi, 1.5))
circ.plot(x)
}
