% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circtools_060226.R
\name{circ.rayleigh}
\alias{circ.rayleigh}
\title{Rayleigh test}
\usage{
circ.rayleigh(x, w = NULL, axial = FALSE, nperm = 0, seed = NULL)
}
\arguments{
\item{x}{Numeric vector of angles (in degrees)}

\item{w}{Optional weights}

\item{axial}{Logical. if TRUE treat data as axial (0–180°)}
}
\description{
Conducts a test of directedness based on inputs. If no weights and axial=F, then the regular Rayleigh test is conducted.
If supplied with weights, the Moore's modified rayleigh test is conducted.
If axial=T, angles are doubled, before conducting test based on weights.
}
