% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circtools_261025.R
\name{circ.stats}
\alias{circ.stats}
\title{Calculate mean statistics}
\usage{
circ.stats(
  x,
  w = NULL,
  type = c("angles", "axial"),
  norm = c("sum", "n"),
  units = c("deg", "rad")
)
}
\arguments{
\item{type}{Type of angles supplied. Default: "angles"}

\item{norm}{Method for normalization of vector length. "sum" = sum(weights)^2 / sum(weights^2). "n" = number of observations}

\item{units}{Unit for your angles. c("deg", "rad")}

\item{angles}{Angles}

\item{weights}{Weights (r-vector lengths)}
}
\description{
Calculate mean direction and vector length. Function supports individual weighting of angles as well as axial data.
When supplied with weights, the function calculates the weighted mean direction and vector length based on an effective sample size of n = sum(w)^2 / sum(w^2).
Argument norm="n", will use the number of observations as the sample size.
}
