% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{circ2genome}
\alias{circ2genome}
\title{Transform from intra-circ coordinates to the genome ones}
\usage{
circ2genome(x, upExon, downExon)
}
\arguments{
\item{x}{a IRanges object}

\item{upexon}{a GRanges object}

\item{downexon}{a GRanges object}
}
\description{
Works only within two adjacent splice junction exons termed
upstream and downatream ones.
}
