% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circtoolsstats_160925.R
\name{circ_MMRT}
\alias{circ_MMRT}
\title{Moores modified Rayleigh Test}
\usage{
circ_MMRT(x, y)
}
\arguments{
\item{x}{Angles in degrees}

\item{y}{r-values (vector lengths)}
}
\description{
Conducts the Moores modified Rayleigh test for weighted data.
}
