% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circtoolsstats_160925.R
\name{circ_rho_zar}
\alias{circ_rho_zar}
\title{Calculate a weighted mean vector length}
\usage{
circ_rho_zar(a, r)
}
\arguments{
\item{a}{A vector of angles in radians}

\item{r}{A vector of r-values (vector lengths)}
}
\description{
Calculate a weighted mean vector length
}
