% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circular_distance_reference}
\alias{circular_distance_reference}
\title{Calculate absolute circular distance between one query value and a vector with one or many values given a fixed period.}
\usage{
circular_distance_reference(query, reference, period = 24)
}
\arguments{
\item{query}{A numeric value}

\item{reference}{A vector with numeric values}

\item{period}{A period for the circle}
}
\description{
Calculate absolute circular distance between one query value and a vector with one or many values given a fixed period.
}
