% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circular_intersect}
\alias{circular_intersect}
\title{Circular-aware intersection of a query value with a time window.}
\usage{
circular_intersect(query, window)
}
\arguments{
\item{query}{The query value.}

\item{window}{A vector of two values representing the lower and upper bounds of the time window.}
}
\description{
Intersect a single timepoint with a timepoint window in a circular-aware fashion given a period,
returning TRUE or FALSE if there is an intersection or not. The lower limit of the window is inclusive,
the upper one is not. Meaning, a query of 2 or a window of 2-4 would return TRUE but for a window 1-2 would
return FALSE.
}
\examples{
query <- c(2)
window1 <- c(0, 3) # TRUE
window2 <- c(23, 3) # TRUE
window3 <- c(3, 5) # FALSE
window4 <- c(23, 2) # FALSE

circular_intersect(query, window1)
circular_intersect(query, window2)
circular_intersect(query, window3)
circular_intersect(query, window4)

}
