% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circtoolsstats_160925.R
\name{confidence_interval_weighted}
\alias{confidence_interval_weighted}
\title{Calculate the weighted confidence intervals}
\usage{
confidence_interval_weighted(
  x,
  conf.level = 0.95,
  w = NULL,
  axial = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{x}{A vector of angles in degrees}

\item{conf.level}{The confidence level to calculate. Default: 0.95}

\item{w}{A vector of r-values (vector lengths)}
}
\description{
Calculate the confidence intervals based on the weighted mean vector length.
}
