% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{decipher2iranges}
\alias{decipher2iranges}
\title{Creates  IRanges objects with metadata on
\itemize{
\item seqId
\item productSize
\item type: \link{'forward', 'reverse'}
\item seq: the primer sequence.
@param primers is a list of results from \link\link{DECIPHER}{DesignPrimers}
@return  a list with an IRanges item for every primer pair
}}
\usage{
decipher2iranges(primers)
}
\description{
Creates  IRanges objects with metadata on
\itemize{
\item seqId
\item productSize
\item type: \link{'forward', 'reverse'}
\item seq: the primer sequence.
@param primers is a list of results from \link\link{DECIPHER}{DesignPrimers}
@return  a list with an IRanges item for every primer pair
}
}
