% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{designPrimers}
\alias{designPrimers}
\title{Design primers}
\usage{
designPrimers(exSeq, db, bsg, opts = list())
}
\arguments{
\item{exSeq}{a GRanges list with GRanges of exons on the splice junction.
Every item represents an individual splice junction.}

\item{db}{}

\item{bsg}{}
}
\description{
Design primers
}
