% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circtoolsstats_011025.R
\name{get.mean.data}
\alias{get.mean.data}
\title{Circular data structure}
\usage{
get.mean.data(
  path = NULL,
  channel = NULL,
  phases = NULL,
  phaselength = NULL,
  axial = NULL,
  samplingrate = 200,
  all = TRUE,
  names = F
)
}
\arguments{
\item{path}{Directory path for the folder to analyse. If left empty, a pop up appears to browse for the folder}

\item{channel}{Which encoder channel your data is stored in (1:4)}

\item{phases}{How many phases in your experiment}

\item{phaselength}{The length of each phase. In seconds}

\item{axial}{List of phases that should be analysed axially. E.g. c(1,2) analyses phase 1 and 2 axially}

\item{samplingrate}{Sampling rate set on your encoder. In milliseconds. Default: 200}

\item{all}{Logical. Calculate statistics for all .txt files, or only the ones that finish your set number of phases.}

\item{names}{Logical. Include specific info from .txt file name. If TRUE, file name should be named with the following info: "EXP_DATE_ID.txt". Lengths of Exp, date and ID can differ, but must be separated with "_".}
}
\description{
Uses .txt outputs from a US Digital Encoder to calculate mean statistics for group analysis.
Calculates mean angles, vector lengths and computes the rayleigh test for all .txt files in a folder.
}
