% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{getCircCoords}
\alias{getCircCoords}
\title{Create circular RNA coordinates GRanges object}
\usage{
getCircCoords(chr, start, end, strand, ids = NULL)
}
\arguments{
\item{chr}{}

\item{start}{}

\item{end}{}

\item{strand}{}

\item{ids}{}
}
\value{
a GRanges object
}
\description{
Create circular RNA coordinates GRanges object
}
