% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{getCircSeqFromList}
\alias{getCircSeqFromList}
\title{Generate splice junction sequencies for a list of circs}
\usage{
getCircSeqFromList(exSeqList)
}
\arguments{
\item{exSeqList}{}
}
\description{
Generate splice junction sequencies for a list of circs
}
