% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circDataEns.R
\name{getExonSeqs}
\alias{getExonSeqs}
\title{Retrieve sequences of the exons, which have common start or end with
the splice junction coordinates provided in circData}
\usage{
getExonSeqs(circData, bsg, faFile, type = c("all", "shortest", "longest"))
}
\arguments{
\item{circData}{a \code{CircData} object}

\item{bsg}{a \code{BSgenome} entity}

\item{faFile}{a \code{\link{FaFile}} object}

\item{type}{a character: whether to include all, only the shortest or
the longest exons, which can form the splice junction}
}
\value{
GRangesList object with a record for every circular id
defined in the \code{circData} argument. The metadata in the list items
include:
\itemize{
\item \code{exon_id} and \code{gene_id}
\item \code{seq}, a DNAStringSet object
\item \code{side}, a character string ("left" or "right")
\item \code{sjId}, a character string derived from the circData object
}
}
\description{
Retrieve sequences of the exons, which have common start or end with
the splice junction coordinates provided in circData
}
