% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circDataEns.R
\name{getSjExons}
\alias{getSjExons}
\title{Get coordinates of splice junction exons}
\usage{
getSjExons(db, circsGR, filter = list())
}
\arguments{
\item{db}{an ensembldb object}

\item{circsGR}{a GRanges of circular splice junctions. 
Must have `CIRCID` column with the identifiers of the circular splice 
juctions.}

\item{filter}{a filter to use for exons output}
}
\value{
a list with `leftSide` and `rightSide` GRanges of the 
intersecting exons.
}
\description{
Get coordinates of splice junction exons
}
