% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{getTxOfSJExons}
\alias{getTxOfSJExons}
\title{Get transcript ids which relate to the circs}
\usage{
getTxOfSJExons(db, exSeq, whichExons = c("any", "both"))
}
\arguments{
\item{db}{ensembldb object}

\item{exSeq}{result of getExonSeqs}

\item{whichExons}{a character:
\itemize{
\item 'any': all transcripts which intersect whether with the left or right
splice junction exons
\item 'both': only those, which intersect with the both, left and right exons.
}}
}
\value{
a list of transcripts ids by circ
}
\description{
Get transcript ids which relate to the circs
}
