% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{getTxSeqs}
\alias{getTxSeqs}
\title{Retirieve transcript seqs for the genes of circs}
\usage{
getTxSeqs(db, bsg, exSeq)
}
\arguments{
\item{db}{ensembldb object}

\item{bsg}{BSGenome or FaFile object
(see \code{x} argument in \code{\link{extractTranscriptSeqs}})}

\item{exSeq}{the result of \code{\link{getExonSeqs}} function}
}
\value{
a list by transcripts with their sequences (\code{\link[Biostrings]{DNAString}})
}
\description{
Retirieve transcript seqs for the genes of circs
}
