% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_data}
\alias{make_data}
\title{Simulate cosinor data}
\usage{
make_data(
  time,
  amplitude = 1,
  acrophase = 1,
  period = 24,
  mesor = 0,
  error_sd = 0.1,
  use_seed = 1
)
}
\arguments{
\item{time}{A numeric vector with time information}

\item{amplitude}{The amplitude of the cosinor curve}

\item{acrophase}{The acrophase of the cosinor curve}

\item{period}{The period of the time series}

\item{mesor}{The mesor of the cosinor curve}

\item{error_sd}{An error value to simulate noise}

\item{use_seed}{A random seed to make error_sd reproducible}
}
\description{
Accept numeric timepoints and cosinor parameters and returns a data.frame with time and values.
}
\examples{
tt <- seq(0, 23.9, by = .1)
data <- make_data(time = tt, acrophase = 5)
plot(data$time, data$value, pch = 20)
abline(v = 5)

}
