% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_design}
\alias{make_design}
\title{Make a design matrix using a cosinor model.}
\usage{
make_design(time, period = 24)
}
\arguments{
\item{time}{A numeric vector with timepoints}

\item{period}{The period of the time series}
}
\description{
Make a design matrix using a cosinor model.
}
