% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circDataEns.R
\name{plotCirc}
\alias{plotCirc}
\title{Plot a transcript scheme}
\usage{
plotCirc(circIds, circGenes, circData)
}
\arguments{
\item{circIds}{}

\item{circGenes}{}

\item{circData}{}
}
\description{
Plot a transcript scheme
}
