% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-segments.R
\name{plotRanges}
\alias{plotRanges}
\title{Plots segments for a list of intervals}
\usage{
plotRanges(ids, starts, ends, seg_width, min_width = 0, xlim = range(starts,
  ends))
}
\arguments{
\item{ids}{a character vector}

\item{starts}{a numeric vector}

\item{ends}{a numerica vector}

\item{seg_width}{a segment size (height) in inches, so it can be the same
for several subplots}

\item{min_width}{a minimal segment width in inches}

\item{xlim}{a range of interval coordinates on the plot. Used for alignment
of features at multiple plots}
}
\value{
Used for its side effect. Plots segments corresponding to given 
intervals.
}
\description{
Plots segments for a list of intervals
}
