% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-segments.R
\name{plotTranscripts}
\alias{plotTranscripts}
\title{Plots isoforms structure, primer position and isoform counts if provided}
\usage{
plotTranscripts(exons, counts = NULL, primers = NULL, circs = NULL,
  minAspectRatio = 0.2, opts = list())
}
\arguments{
\item{exons}{an interval data.frame}

\item{counts}{an interval data.frame}

\item{primers}{an interval data.frame}

\item{circs}{a counts data.frame}

\item{minAspectRatio}{a minimal ratio of a segment height to its width}

\item{opts}{an option list}
}
\value{
Used for its side effects. Plots intervals for exons,
primers and transcript counts if provided.
}
\description{
Plots isoforms structure, primer position and isoform counts if provided
}
\details{
all interval data.frames must have start and end fields.
\code{exons} is assumed to have \code{strand} (character) and \code{tx_id} fields.
\code{primers} must have \code{id} to distinguish forward, reverse primers for
different circular transcripts.
\code{cirsc} must contain sjId field.
\code{counts} has id and count fields.
}
