% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rad2period}
\alias{rad2period}
\title{Convert radians hours to period hours.}
\usage{
rad2period(x, period = 24)
}
\arguments{
\item{x}{a radians time value, from 0 to 2*pi.}

\item{period}{The numeric period.}
}
\description{
Convert radians hours to period hours.
}
