% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-junctions.R
\name{reportCircs}
\alias{reportCircs}
\title{Write an HTML report for circular RNA candidates}
\usage{
reportCircs(exSeq, file)
}
\arguments{
\item{exSeq}{an object created by \code{\link{getExSeq}}}

\item{file}{a report file name}
}
\description{
Write an HTML report for circular RNA candidates
}
