% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circDataEns.R
\name{CircData}
\alias{CircData}
\title{Create a CircData object}
\usage{
CircData(db, circCoords)
}
\arguments{
\item{db}{an ensembldb object}

\item{circCoords}{an IRanges object}
}
\value{
a CircData object
}
\description{
Create a CircData object
}
