% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{designPrimers}
\alias{designPrimers}
\title{Design primers}
\usage{
designPrimers(exSeq, db, bsg, opts = list(lengthRange = c(0, Inf)))
}
\arguments{
\item{exSeq}{a GRanges list with GRanges of exons on the splice junction.
Every item represents an individual splice junction.}

\item{db}{an ensembldb object}

\item{bsg}{a BSgenome object}

\item{opts}{a named list with the design options (see details).}
}
\value{
a list with \code{primers} and \code{product} items.
}
\description{
Design primers
}
\details{
\code{opts} can include parameters for \code{\link[DECIPHER]{designPrimers}} and
a vector lengthRange for further filtering by the product length
(default = c(0,Inf))
}
