% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circtoolsstats_160925.R
\name{get.mean.data}
\alias{get.mean.data}
\title{Circular data structure}
\usage{
get.mean.data(
  channel = NULL,
  phases = NULL,
  phaselength = NULL,
  samplingrate = 200,
  all = TRUE,
  names = F
)
}
\arguments{
\item{channel}{Which encoder channel your data is stored in (1:4)}

\item{phases}{How many phases in your experiment}

\item{phaselength}{The length of each phase. In seconds}

\item{samplingrate}{Sampling rate set on your encoder. In milliseconds. Default: 200}

\item{all}{Logical. Calculate statistics for all .txt files, or only the ones that finish your set number of phases.}

\item{names}{Logical. Include specific info from .txt file name. If TRUE, file name should be named with the following info: "EXP_DATE_ID.txt". Lengths of Exp, date and ID can differ, but must be separated with "_".}
}
\description{
Calculates mean angles, vector lengths and computes the rayleigh test for all .txt files in a folder.
}
