% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circDataEns.R
\name{getExonSeqs}
\alias{getExonSeqs}
\title{Retrieve sequences of the exons, which have common start or end with 
the splice junction coordinates provided in circData}
\usage{
getExonSeqs(circData, bsg, type = c("all", "shortest", "longest"))
}
\arguments{
\item{circData}{a `CircData` object}

\item{bsg}{a `BSgenome` entity}
}
\value{
a list of `leftSide` and `rightSide` exons
}
\description{
Retrieve sequences of the exons, which have common start or end with 
the splice junction coordinates provided in circData
}
\examples{
\dontrun{

}
}
