% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circDataEns.R
\name{plotCirc}
\alias{plotCirc}
\title{Plot a transcript scheme}
\usage{
plotCirc(sjIds, circGenes, circData, primers = NULL, counts = NULL,
  countThres = -Inf, opts = NULL)
}
\arguments{
\item{sjIds}{a character identifiers for the circular transcript
to be plotted.}

\item{circGenes}{a character vector or list with the gene id's, for which
circular transcripts must be plotted}

\item{circData}{an object returned by CircData}

\item{primers}{a data.frame or an IRanges object for the primers}

\item{counts}{a data.frame with an id column (tx_id) and corresponding
read counts.}

\item{countThres}{the transcripts with the read counts exceeding the
threshold value will be plotted (default: 0)}

\item{opts}{a list with options for the \code{\link{plotTranscripts}}
opts argument.}
}
\description{
Plot a transcript scheme
}
