% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circDataEns.R
\name{plotCirc}
\alias{plotCirc}
\title{Plot a transcript scheme}
\usage{
plotCirc(circIds, circGenes, circData, primers = NULL, counts = NULL)
}
\arguments{
\item{circIds}{a character identifiers for the circular transcript 
to be plotted.}

\item{circData}{an object returned by CircData}

\item{primers}{a data.frame or an IRanges object for the primers}

\item{counts}{a data.frame with an id column (tx_id) and corresponding
read counts.}

\item{circGene}{a character identifier for a gene, which transcripts are to 
be plotted. If `circIds` is not set, all circular transcripts defined in
`circData` will be used. The identifier is identical to the GENEID
record in the EnsDb or TxDB objects.}
}
\description{
Plot a transcript scheme
}
