% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-segments.R
\name{plotRanges}
\alias{plotRanges}
\title{Plots segments for a list of intervals}
\usage{
plotRanges(ids, starts, ends, segmentSize, minWidth = 0,
  xlim = range(starts, ends), ylim = c(0, 1 + length(levels(ids))), opts)
}
\arguments{
\item{ids}{a character vector}

\item{starts}{a numeric vector}

\item{ends}{a numerica vector}

\item{segmentSize}{a segment size (height) in inches, so it can be the same
for several subplots}

\item{minWidth}{a minimal segment width in inches}

\item{xlim}{a range of interval coordinates on the plot. Used for alignment
of features at multiple plots}

\item{ylim}{a ylim range}

\item{opts}{a list of plotting parameters:
- col, the color of the segments (default: "dodgerblue4")
- connect, a logical. If the segments must be connected with lines 
(default: TRUE).}
}
\value{
Used for its side effect. Plots segments corresponding to given 
intervals.
}
\description{
Plots segments for a list of intervals
}
