% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-segments.R
\name{plotTranscripts}
\alias{plotTranscripts}
\title{Plots isoforms structure, primer position and isoform counts if provided}
\usage{
plotTranscripts(exons, counts = NULL, primers = NULL, circs = NULL,
  minRatio = 0.2, opts = list())
}
\arguments{
\item{exons}{an interval data.frame}

\item{counts}{an interval data.frame}

\item{primers}{an interval data.frame}

\item{circs}{a counts data.frame}

\item{minRatio}{a minimal aspect ratio for plotted segments (height/width)}

\item{opts}{an option list}
}
\value{
Used for its side effects. Plots intervals for exons,
primers and transcript counts if provided.
}
\description{
Plots isoforms structure, primer position and isoform counts if provided
}
